requirements.txt
dj-database-url==0.4.0
Django==1.9.2
djangorestframework==3.3.3
gunicorn==19.4.5
httpie==0.9.3
psycopg2==2.6.1
Pygments==2.1.3
requests==2.9.1
virtualenv==15.0.1
whitenoise==2.0.6
